refactor into internal/arrow
