+++ exporter/otlpexporter/internal/arrow/config.go
@@ -1,0 +1,33 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package arrow
+
+import "fmt"
+
+// ArrowSettings includes whether Arrow is enabled and the number of
+// concurrent Arrow streams.
+type ArrowSettings struct {
+	Enabled    bool `mapstructure:"enabled"`
+	NumStreams int  `mapstructure:"num_streams"`
+}
+
+// Validate returns an error when the number of streams is less than 1.
+func (cfg *ArrowSettings) Validate() error {
+	if cfg.NumStreams < 1 {
+		return fmt.Errorf("stream count should be > 0: %d", cfg.NumStreams)
+	}
+
+	return nil
+}
